--- a/torchtune/training/checkpointing/__init__.py
+++ b/torchtune/training/checkpointing/__init__.py
@@ -5,6 +5,11 @@
 # LICENSE file in the root directory of this source tree.
 from typing import Union
 
+from torchtune.training.checkpointing._checkpoint_client import (
+    CheckpointClient,
+    TrainingProgress,
+)
+
 from torchtune.training.checkpointing._checkpointer import (
     DistributedCheckpointer,
     FullModelHFCheckpointer,
@@ -36,6 +41,7 @@ Checkpointer = Union[
 ]
 
 __all__ = [
+    "CheckpointClient",
     "FullModelHFCheckpointer",
     "FullModelMetaCheckpointer",
     "FullModelTorchTuneCheckpointer",
@@ -55,4 +61,5 @@ __all__ = [
     "STEPS_KEY",
     "TOTAL_EPOCHS_KEY",
     "FormattedCheckpointFiles",
+    "TrainingProgress",
 ]