--- a/torchtune/training/checkpointing/__init__.py
+++ b/torchtune/training/checkpointing/__init__.py
@@ -6,6 +6,7 @@
 from typing import Union
 
 from torchtune.training.checkpointing._checkpointer import (
+    DistributedCheckpointer,
     FullModelHFCheckpointer,
     FullModelMetaCheckpointer,
     FullModelTorchTuneCheckpointer,
@@ -28,6 +29,7 @@ from torchtune.training.checkpointing._utils import (
 )
 
 Checkpointer = Union[
+    DistributedCheckpointer,
     FullModelHFCheckpointer,
     FullModelMetaCheckpointer,
     FullModelTorchTuneCheckpointer,
@@ -37,6 +39,7 @@ __all__ = [
     "FullModelHFCheckpointer",
     "FullModelMetaCheckpointer",
     "FullModelTorchTuneCheckpointer",
+    "DistributedCheckpointer",
     "ModelType",
     "Checkpointer",
     "update_state_dict_for_classifier",