@felipemello1 2024-12-09
@saumishr, there were considerable changed made to the output path of the files. Can you double check that we can still resume from checkpoint even if we are using dcp? https://github.com/pytorch/torchtune/pull/2074

Also, please check the folder structure. The assumption is that the use can do `cp path_with_ckpt path_with_cfgs`, and this is ready to upload o HF. But if there are 50 files there related to dcp, this will look a bit ugly.

I think that we also need to have a clear solution about going from intermediate ckpt -> hf_format. Like we discussed, many of our users run eval on the intermediate ckpts.

cc: @joecummings