@joecummings 2024-12-12
> @joecummings There are no tests failing. A CI/CD job is complaining that the diff needs to be updated with the latest. I'm validating the async checkpoints with the recent refactor @felipemello1 had done. I will notify here once the validation is done and the PR can be merged. Thank you!

Once this validation is confirmed, I'm happy to merge this!